[REFACTOR] : .png -> .webp 변경
